+++ plugins/auth-backend-module-microsoft-provider/config.d.ts
@@ -21,11 +21,12 @@ export interface Config {
       microsoft?: {
         [authEnv: string]: {
           clientId: string;
+          tenantId: string;
           /**
            * @visibility secret
            */
-          tenantId: string;
           clientSecret: string;
+          domainHint?: string;
           callbackUrl?: string;
         };
       };