+++ plugins/azure-devops/src/hooks/index.ts
@@ -16,8 +16,8 @@
 
 export * from './useAllTeams';
 export * from './useDashboardPullRequests';
-export * from './useProjectRepoFromEntity';
 export * from './usePullRequests';
 export * from './useRepoBuilds';
 export * from './useUserEmail';
 export * from './useUserTeamIds';
+export * from './useReadme';