+++ plugins/scaffolder/src/api.ts
@@ -38,9 +38,7 @@ import {
   ScaffolderDryRunOptions,
   ScaffolderDryRunResponse,
   TemplateParameterSchema,
-  ScaffolderStep,
 } from '@backstage/plugin-scaffolder-react';
-import { TaskStep } from '@backstage/plugin-scaffolder-common';
 
 import queryString from 'qs';
 import { EventSourcePolyfill } from 'event-source-polyfill';
@@ -172,15 +170,11 @@ export class ScaffolderClient implements ScaffolderApi {
 
   async getTask(taskId: string): Promise<ScaffolderTask> {
     const baseUrl = await this.discoveryApi.getBaseUrl('scaffolder');
-    const taskUrl = `${baseUrl}/v2/tasks/${encodeURIComponent(taskId)}`;
+    const url = `${baseUrl}/v2/tasks/${encodeURIComponent(taskId)}`;
 
-    const taskEventsUrl = `${baseUrl}/v2/tasks/${encodeURIComponent(
-      taskId,
-    )}/events`;
-
-    const taskResponse = await this.fetchApi.fetch(taskUrl);
-    if (!taskResponse.ok) {
-      throw await ResponseError.fromResponse(taskResponse);
+    const response = await this.fetchApi.fetch(url);
+    if (!response.ok) {
+      throw await ResponseError.fromResponse(response);
     }
 
     const taskEventsResponse = await this.fetchApi.fetch(taskEventsUrl);