+++ plugins/catalog-graphql/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-catalog-graphql
 
+## 0.4.1-next.0
+
+### Patch Changes
+
+- 6c2b872153: Add official support for React 18.
+- Updated dependencies
+  - @backstage/catalog-model@1.4.3
+  - @backstage/config@1.1.1
+  - @backstage/types@1.1.1
+
 ## 0.4.0
 
 ### Minor Changes