+++ plugins/lighthouse-backend/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/plugin-lighthouse-backend
 
+## 0.3.4-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.19.9-next.0
+  - @backstage/backend-plugin-api@0.6.7-next.0
+  - @backstage/backend-tasks@0.5.12-next.0
+  - @backstage/catalog-client@1.4.5
+  - @backstage/catalog-model@1.4.3
+  - @backstage/config@1.1.1
+  - @backstage/types@1.1.1
+  - @backstage/plugin-catalog-node@1.4.8-next.0
+  - @backstage/plugin-lighthouse-common@0.1.4
+
 ## 0.3.3
 
 ### Patch Changes